requirements.txt
ajax==3.0.0
appdirs==1.4.4
asgiref==3.3.4
beautifultable==1.0.1
black==20.8b1
certifi==2020.12.5
chromedriver-autoinstaller==0.2.2
click==7.1.2
colorama==0.4.4
decorator==5.0.7
Django==3.2.2
django-appconf==1.0.4
django-debug-toolbar==3.2.1
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
drf-nested-routers==0.93.3
flake8==3.8.4
flake8-html==0.4.1
future==0.18.2
geographiclib==1.50
geopy==2.1.0
h11==0.12.0
httpcore==0.12.3
httpx==0.17.1
idna==3.1
ImageHash==4.2.0
importlib-metadata==3.7.0
javascript==2015.32.59
Jinja2==2.11.3
mailcap-fix==1.0.1
Markdown==3.3.4
MarkupSafe==1.1.1
mccabe==0.6.1
mime==0.1.0
mypy-extensions==0.4.3
numpy==1.20.1
pandas==1.2.2
pathspec==0.8.1
Pillow==8.1.2
psycopg2==2.8.6
pycodestyle==2.6.0
pyflakes==2.2.0
Pygments==2.8.0
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.1.1
regex==2020.11.13
rfc3986==1.4.0
routers==0.10.1
scipy==1.6.1
selenium==3.141.0
selenium-wire==2.1.2
six==1.15.0
sniffio==1.2.0
sqlparse==0.4.1
termcolor==1.1.0
tinydb==4.4.0
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.4
wcwidth==0.2.5
zipp==3.4.0
